requirements.txt
pandas
openpyxl
numpy
matplotlib
seaborn
jupyter
tqdm
nltk
spacy
gensim
scikit-learn
torch
transformers
tensorflow
tensorflow-hub
tensorboard
streamlit
pyLDAvis
lime
shap
faiss-cpu
sentence-transformers
langchain
langchain-community
chromadb
datasets
accelerate
ipywidgets